+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "target": "es5",
     "module": "commonjs",
-    "lib": ["ES2018"],
+    "lib": ["ES2020"],
     "outDir": "dist",
     "declaration": true,
     "strict": true,
